requirements.txt
arrow==0.12.1
beautifulsoup4==4.6.1
certifi==2018.4.16
chardet==3.0.4
Jinja2==2.10
MarkupSafe==1.0
pkg-resources==0.0.0
psycopg2-binary==2.7.5
python-dateutil==2.7.3
pytz==2018.5
selenium==3.14.1
six==1.11.0
tzlocal==1.5.1
urllib3==1.23
